last update

+++ src/main/java/com/example/paysvc/repository/DebtRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DebtRepository extends JpaRepository<DebtEntity, Long> {
 
-    DebtResponse findByAccountCode(int accountCode);
+    DebtResponse findByAccountCode(long accountCode);
 }
